src/SoftTmi{,Fast}Interface.h: Change type of delayMicros to uint8_t, saving 4 bytes of flash on AVR

+++ src/ace_segment/hw/SoftTmiFastInterface.h
@@ -67,7 +67,7 @@ namespace ace_segment {
 template <
     uint8_t T_CLK_PIN,
     uint8_t T_DIO_PIN,
-    uint16_t T_DELAY_MICROS
+    uint8_t T_DELAY_MICROS
 >
 class SoftTmiFastInterface {
   public:
